product detail get api done
