+++ ecommerce_project/myapp/views/__init__.py
@@ -1,3 +1,4 @@
 from .user_views import *
 from .other_views import *
 from .product_related_views import *
+from .product_views import *
