+++ ecommerce_project/urls.py
@@ -29,6 +29,7 @@ urlpatterns = [
     path('addresses/', views.AddressListNCreateAPIView.as_view()),
     path('categories/', views.CategoryListNCreateAPIView.as_view()),
     path('products/', views.ProductListNCreateAPIView.as_view()),
+    path('products/<int:pk>/', views.ProductDetailUpdateDeleteAPIView.as_view()),
     path('reviews/', views.ReviewListNCreateAPIView.as_view()),
     path('orders/', views.OrderListNCreateAPIView.as_view()),
     path('order-lines/', views.OrderLineListNCreateAPIView.as_view()),
